backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
Pillow==10.2.0
requests
celery
redis
channels
daphne
channels-redis
whitenoise
langchain-google-genai
langgraph